+++ Cargo.toml
@@ -1,12 +1,22 @@
 [package]
-name = "grafana-alerting-email"
+name = "alerting_email"
 version = "0.0.1"
 authors = ["Bodo Hugo Barwich <b.barwich@hotmail.com>"]
 edition = "2018"
 
+[[bin]]
+name = "grafana-alerting-email"
+path = "src/main.rs"
+
+[lib]
+name = "alerting_email"
+path = "src/lib.rs"
+
 [dependencies]
 actix = "=0.10"
 actix-web = "3"
+actix-rt = "*"
+actix-test = "*"
 tokio = "=0.1.14"
 futures = "=0.1.29"
 futures-core = "=0.3.11"
